fix(gosfx): SpecLibrary::Save writes 32-bit count (x64 port bug) + mc2fx slice 2
SpecLibrary::Save emitted the effect count via `*stream << GetLength()`, where
DynamicArrayOf::GetLength() returns size_t — 8 bytes on x64. The loader reads it
as `unsigned len` (4 bytes). Original 32-bit MC2 had size_t==unsigned so this was
symmetric; the 64-bit port silently wrote a stray zero high-dword after the count,
desyncing every following spec and producing an UNLOADABLE blob. Cast to unsigned
to match the read side. This was dormant — nothing in the shipping game/editor
calls EffectLibrary::Save, so the rot went unnoticed until mc2fx exercised it.

mc2fx slice 2:
- mc2fx_core.{h,cpp}: extracted reusable headless load/dump/save core
  (initEngineHeadless / loadBlob / dumpCatalogJson / saveBlob) for previewer reuse.
- `rebuild <in.fx> <out.fx>` subcommand: Load -> Save -> byte-compare round-trip gate.
- saveBlob reserves the DynamicMemoryStream up front (workaround for a separate
  latent engine bug: DynamicMemoryStream::WriteBytes(void*,DWORD) fails to override
  base WriteBytes(void*,size_t) on x64, so the growing path is dead — flagged for
  separate fix; not touched here).
- MC2FX_ASAN cmake option (off by default) for sanitizer builds.

Round-trip now: out.fx RELOADS, dump(in)==dump(out) JSON-identical (904 effects).
Remaining: not byte-exact (float canonicalization @417, stable) and not yet
fixed-point (nondeterministic uninit bytes in some curve/state Save — next).

+++ mclib/particles/spec_library.cpp
@@ -70,7 +70,13 @@ void SpecLibrary::Load(Stuff::MemoryStream* stream)
 void SpecLibrary::Save(Stuff::MemoryStream* stream)
 {
     gosFX::WriteGFXVersion(stream);
-    *stream << m_effects.GetLength();
+    // 64-bit port fix: GetLength() returns size_t (8 bytes on x64), but the
+    // read side (Load: `*stream >> len` with `unsigned len`) consumes only 4.
+    // The original 32-bit MC2 had size_t==unsigned so this was symmetric; on
+    // x64 the unchecked 8-byte write injected a stray high-dword of zeros after
+    // the count, desyncing every following spec and producing an unloadable
+    // blob. Cast to the same 32-bit width the loader reads.
+    *stream << static_cast<unsigned>(m_effects.GetLength());
     for (unsigned i = 0; i < m_effects.GetLength(); ++i) {
         Check_Object(m_effects[i]);
         m_effects[i]->Save(stream);
